hugging_face_moderation_service/requirements.txt
fastapi==0.115.12
httpx==0.28.1
pydantic==2.11.3
pydantic_settings==2.9.1
python-dotenv==1.1.0
Requests==2.32.3
uvicorn==0.34.2
